fix(check): glance is the shared named-permits cell
the audited-card test pinned permits declared. MCP's green answer
said the same opaque word. both now call EffectivePermits::glance so
the CLI card, the MCP contract and --json cannot drift.
