+++ crates/nika-display/src/check_render/permits_glance.rs
@@ -6,57 +6,13 @@
 //! already named the grants; the default card said `declared`.
 
 use nika_check::CheckReport;
-use nika_schema::types::{ExecPermit, Permits};
 
-/// Compact grant list for the audited card. Absent is `none`; an
-/// explicit empty block is `{}` (the legal zero, not the undeclared
-/// one). Cells join on spaces so the outer ` · ` separators stay one
-/// field.
+/// Compact grant list for the audited card. Delegates to
+/// [`nika_check::EffectivePermits::glance`] so MCP and the CLI card
+/// cannot drift.
 #[must_use]
 pub(super) fn permits_glance(report: &CheckReport) -> String {
-    match report.permits.declared.as_ref() {
-        None => "none".to_owned(),
-        Some(p) => format_grants(p),
-    }
-}
-
-fn format_grants(p: &Permits) -> String {
-    let mut cells = Vec::new();
-    match &p.exec {
-        Some(ExecPermit::Any) => cells.push("exec:any".to_owned()),
-        Some(ExecPermit::Programs(ps)) if !ps.is_empty() => {
-            cells.push(format!("exec:{}", ps.join(",")));
-        }
-        _ => {}
-    }
-    if let Some(tools) = &p.tools
-        && !tools.is_empty()
-    {
-        cells.push(format!("tools:{}", tools.join(",")));
-    }
-    if let Some(fs) = &p.fs {
-        if !fs.read.is_empty() {
-            cells.push(format!("read:{}", fs.read.join(",")));
-        }
-        if !fs.write.is_empty() {
-            cells.push(format!("write:{}", fs.write.join(",")));
-        }
-    }
-    if let Some(net) = &p.net
-        && !net.http.is_empty()
-    {
-        cells.push(format!("http:{}", net.http.join(",")));
-    }
-    if let Some(env) = &p.env
-        && !env.is_empty()
-    {
-        cells.push(format!("env:{}", env.join(",")));
-    }
-    if cells.is_empty() {
-        "{}".to_owned()
-    } else {
-        cells.join(" ")
-    }
+    report.permits.glance()
 }
 
 #[cfg(test)]
